+++ pytsite/core/util.py
@@ -180,3 +180,15 @@ def dict_cleanup(inp: dict) -> dict:
             r[k] = v
 
     return r
+
+
+def nav_link(href: str, anchor: str, **kwargs: dict) -> str:
+    """Generate 'navigation' link.
+    """
+    from . import html, router
+
+    li = html.Li()
+    if href != '#' and router.url(href, strip_query=True) == router.current_url(strip_query=True):
+        li.set_attr('cls', 'active')
+
+    return str(li.append(html.A(anchor, href=href, **kwargs)))